Allow commands to be input from a floating window
This adds a new `command_floating_window` option. When set to true, gp
will use a floating window + scratch buffer to read the command. The
map to accept the input is the same as the one to send a message in the
chat window (ctrl+g, ctrl+g by default).